requirements.txt
APScheduler==3.7.0
asn1crypto==1.4.0
attrs==21.2.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.6
chardet==4.0.0
configparser==4.0.2
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==3.3.2
enum34==1.1.10
funcsigs==1.0.2
futures==3.3.0
hash-ring==1.3.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==2.1.1
incremental==21.3.0
ipaddress==1.0.23
kazoo==2.8.0
pathlib2==2.3.6
ply==3.11
py-dateutil==2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.10.1
PyHamcrest==1.10.1
pyOpenSSL==20.0.1
pysmi==0.3.4
pysnmp==4.4.12
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
scandir==1.10.0
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.23
thrift==0.13.0
Twisted==20.3.0
typing==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
ipp==1.2.0
zope.interface==5.4.0
